requirements.txt
streamlit
networkx
gitpython